refactor(foundation-demo): change file names of daff-footer to footer
change file names of daff-footer to footer
