+++ apps/foundation-demo/src/app/design/molecules/footer/components/footer/footer.component.scss
