+++ apps/foundation-demo/src/app/misc/footer/footer.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { FooterComponent } from './footer.component';
-import { DaffFooterModule } from '../../design/molecules/daff-footer/daff-footer.module';
+import { DaffFooterModule } from '../../design/molecules/footer/footer.module';
 
 describe('FooterComponent', () => {
   let component: FooterComponent;
